Remove legacy files, rename directory, remove requirement for sse4.2
