+++ NOTICE.txt
@@ -52,10 +52,6 @@ This product includes software from the ClickHouse project
  * Copyright 2016-2022 ClickHouse, Inc.
  * https://github.com/ClickHouse/ClickHouse
 
-This product includes software from the BBHash project
- * Copyright (c) 2015 Guillaume Rizk
- * https://github.com/rizkg/BBHash
-
 This product includes software from the rax project (BSD, 2-clause)
  * Copyright (c) 2017-2019, Salvatore Sanfilippo <antirez at gmail dot com>
  * https://github.com/antirez/rax
