+++ modules/basic/ds/perfect_hash/single_phf_view.h
@@ -13,22 +13,23 @@
  * limitations under the License.
  */
 
-#ifndef MODULES_BASIC_DS_GRAPE_PERFECT_HASH_SINGLE_PHF_VIEW_H_
-#define MODULES_BASIC_DS_GRAPE_PERFECT_HASH_SINGLE_PHF_VIEW_H_
+#ifndef MODULES_BASIC_DS_PERFECT_HASH_SINGLE_PHF_VIEW_H_
+#define MODULES_BASIC_DS_PERFECT_HASH_SINGLE_PHF_VIEW_H_
 
 #include <cstdlib>
 #include <set>
 #include <vector>
 
-#include "basic/ds/grape_perfect_hash/encoders_view.h"
-#include "basic/ds/grape_perfect_hash/ref_vector.h"
+#include "basic/ds/perfect_hash/encoders_view.h"
+#include "basic/ds/perfect_hash/ref_vector.h"
 #include "pthash/pthash.hpp"
 #include "pthash/single_phf.hpp"
 
 #define nssv_CONFIG_SELECT_STRING_VIEW nssv_STRING_VIEW_NONSTD
 #include "string_view/string_view.hpp"
 
-namespace grape_perfect_hash {
+namespace vineyard {
+namespace perfect_hash {
 
 struct murmurhasher {
   typedef pthash::hash64 hash_type;
@@ -181,6 +182,7 @@ struct SinglePHFView {
   ef_sequence_view m_free_slots;
 };
 
-}  // namespace grape_perfect_hash
+}  // namespace perfect_hash
+}  // namespace vineyard
 
-#endif  // MODULES_BASIC_DS_GRAPE_PERFECT_HASH_SINGLE_PHF_VIEW_H_
+#endif  // MODULES_BASIC_DS_PERFECT_HASH_SINGLE_PHF_VIEW_H_
