+++ modules/basic/ds/perfect_hash/string_view_vector.h
@@ -13,8 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#ifndef MODULES_BASIC_DS_GRAPE_PERFECT_HASH_STRING_VIEW_VECTOR_H_
-#define MODULES_BASIC_DS_GRAPE_PERFECT_HASH_STRING_VIEW_VECTOR_H_
+#ifndef MODULES_BASIC_DS_PERFECT_HASH_STRING_VIEW_VECTOR_H_
+#define MODULES_BASIC_DS_PERFECT_HASH_STRING_VIEW_VECTOR_H_
 
 #include <stdlib.h>
 #include <string.h>
@@ -23,11 +23,12 @@ limitations under the License.
 #include <utility>
 #include <vector>
 
-#include "basic/ds/grape_perfect_hash/ref_vector.h"
+#include "basic/ds/perfect_hash/ref_vector.h"
 #define nssv_CONFIG_SELECT_STRING_VIEW nssv_STRING_VIEW_NONSTD
 #include "string_view/string_view.hpp"
 
-namespace grape_perfect_hash {
+namespace vineyard {
+namespace perfect_hash {
 
 class StringViewVector {
  public:
@@ -125,6 +126,7 @@ struct ref_vector<nonstd::string_view> {
   ref_vector<size_t> offsets_;
 };
 
-}  // namespace grape_perfect_hash
+}  // namespace perfect_hash
+}  // namespace vineyard
 
-#endif  // MODULES_BASIC_DS_GRAPE_PERFECT_HASH_STRING_VIEW_VECTOR_H_
+#endif  // MODULES_BASIC_DS_PERFECT_HASH_STRING_VIEW_VECTOR_H_
